enhanced types to atoms
