+++ modules/shared/components/atoms/Button/types/IButton.d.ts
@@ -1,15 +1,18 @@
 import type * as EButton from './EButton';
 
 declare namespace IButton {
-  export interface IProps {
-    children?: string | null;
+  export interface IProps extends IHandlers {
+    children?: string;
     disabled?: boolean;
     variant: EButton.buttonVariantValues;
     size: EButton.buttonSizeValues;
     leftIcon?: boolean;
     rightIcon?: boolean;
     onlyIcon?: boolean;
-    onClick?: () => void;
+  }
+
+  export interface IHandlers {
+    onClick: () => void;
   }
 }
 export { IButton };
