+++ modules/shared/components/atoms/checkbox/types/ICheckbox.d.ts
@@ -1,11 +1,8 @@
 declare namespace ICheckbox {
-  export interface ICheckBoxProps {
-    disabled: boolean;
+  export interface IProps {
+    disabled?: boolean;
     size: 'Default' | 'Small';
     onMockClick?: () => void;
-  }
-  export interface ICheckBoxSvgProps {
-    style: string;
   }
 }
 export { ICheckbox };
