+++ modules/shared/components/molecules/UserInfo/Index.tsx
@@ -7,13 +7,13 @@ import styles from './UserInfo.module.css';
 const UserInfo: FC<IUserInfo.IProps> = ({
   size,
   variant,
-  imgSrc,
+  profile_pic,
   name,
   date,
 }): ReactElement => {
   return (
     <div className={styles['outer-wrapper']}>
-      <Avatar size={size} variant={variant} imgSrc={imgSrc} />
+      <Avatar size={size} variant={variant} profile_pic={profile_pic} />
       <div className={styles['user-wrapper']}>
         <span className={styles.name} data-testid="name">
           {variant === 'anonymous' ? 'Anonymous' : name}
